# Faster Flash writes, both on die and SPI