… speeduino/src/SPIAsEEPROM/winbondflash.h