On die writes got a small change 38 to 34us, SPI flash got a big improvement, average byte write changed from 144 to 93us.

Tested on STM32F407VE board on both cases.
Need to be tested on Teensy.

[FlashAsEEPROMTest.zip](https://github.com/noisymime/speeduino/files/9998938/FlashAsEEPROMTest.zip)
This is @Tjeerdie project used to test the speed, **_this will wipe all the eeprom/flash content!_** **Test on an unused board**