@DeionSi 2023-01-08
I get a few compiler warnings with this when compiling for mega. I know this doesn't apply to mega but I guess it shouldn't generate warnings for it.